Move promise utils to promiseUtils.ts
