+++ src/cardano/CardanoEngine.ts
@@ -20,8 +20,9 @@ import { base16 } from 'rfc4648'
 import { CurrencyEngine } from '../common/CurrencyEngine'
 import { PluginEnvironment } from '../common/innerPlugin'
 import { getRandomDelayMs } from '../common/network'
+import { promiseCast } from '../common/promiseUtils'
 import { trial } from '../common/trial'
-import { cleanTxLogs, getFetchCors, promiseCast } from '../common/utils'
+import { cleanTxLogs, getFetchCors } from '../common/utils'
 import { asStakingTxBody } from './asStakingTx'
 import { CardanoTools } from './CardanoTools'
 import {
