+++ src/ethereum/EthereumNetwork.ts
@@ -3,7 +3,8 @@ import { EdgeTransaction } from 'edge-core-js/types'
 
 import { getRandomDelayMs } from '../common/network'
 import { makePeriodicTask, PeriodicTask } from '../common/periodicTask'
-import { asyncWaterfall, normalizeAddress, promiseAny } from '../common/utils'
+import { asyncWaterfall, promiseAny } from '../common/promiseUtils'
+import { normalizeAddress } from '../common/utils'
 import { WEI_MULTIPLIER } from './ethereumConsts'
 import { EthereumEngine } from './EthereumEngine'
 import { EthereumNetworkInfo } from './ethereumTypes'
