+++ src/fio/FioTools.ts
@@ -16,9 +16,9 @@ import {
 } from 'edge-core-js/types'
 
 import { PluginEnvironment } from '../common/innerPlugin'
+import { asyncWaterfall } from '../common/promiseUtils'
 import { encodeUriCommon, parseUriCommon } from '../common/uriHelpers'
 import {
-  asyncWaterfall,
   getFetchCors,
   getLegacyDenomination,
   mergeDeeply,
